my-swaps component directory moved to custom from swap market
